Removed the "flexible initializer"
Being able to create `Card` objects by passing symbols, ints, strings,
or constants names makes for a friendly DSL, but it brings a lot of
internal complexity. It's not worth it. By restricting the API, we
can also remove a lot of code, to the point that using `Data` objects
makes sense again. Even if the code for `#to_short_s` is a bit
cryptics, I find the final result friendlier.