Introducing Hand
Implementing a hand of cards with a simple Array will be tedious
now that we have to serialize/deserialize it. So let's introduce a
model for that.

Note that I almost put it in the `/app/models` folder, since its logic is
so specific to the application. Ah, the recurring dilema of the `/lib`
folder…