Renamed Card#new_hand to Card#one_of_each_suit
The name is a bit too descriptive, but since we now have `Hand` objects,
and this methods does **not** return one, let's avoid confusions.

(I don't want the method to return a `Hand` object because I don't want
to add coupling; besides, a "hand" is any number of cards, not necessarily
one of each suit – that's just the _initial_ hand. Which might as well
be under the responsibility of `Round` or `Game`. Something to consider,
to be honest…)