Removed #fill_hand but kept the logic
Now, this is an interesting commit. In a previous one, I introduced
`Round#fill_hand`, as an implementation of the rule "Draw two cards from
the draw pile that matches the suit of the current Event card. Place
them into your hand.", which is part of the round sequence. (In Mavrides,
the round sequence is pre-executed for both players when a new round
starts, in `Round#next`.)

This small piece of implementation had been put in its own method
mostly to make is easier to test, since it would require a specific
setup: previous events had to be prepared so for the remaining draw
pile to be in a controlled state suitable for a test.

However, working on it, I realized the logic behind the computation of
the draw pile(s) should be under the `Game` responsibility, therefore
simplifying the test setup, which in turn made the `#fill_hand` method
less relevant. So I moved the significant parts of the test (including
_a bit_ more setup) to the test for `Round#next`.