Updating the logic of Round#next
Drawing events **and** refills for the hand is confusing is the two
things happen at different times, so let's encapsulate everything in
`Round#next`, the designated factory method for a new, complete round.

A new Round created through this method will be in its final state (i.e.
ready to be replaced by a new round):

-   The events have been drawn from the previous hand
-   The new hands are set (without the drawn event, but with 2 new cards)