Changed: a new round starts w/ **Earther**'s msg
Since we're not waiting for a (new) round's message to be received to
create said round, but instead create it _before_ the message is sent,
then the creation of new rounds should be triggered by the reception of
the **Earther**'s message (which is the last one in a round), and not
the Explorer's.

Also note that, because new rounds are generated "earlier", but
fully-formed, I had to remove the `running_game_round_two` fixture,
which wasn't legit (it was empty). Now the game in the fixture is on its
1st round, with the Explorer's message already received, and the
Earther's still waiting to be processed.