Move the event prompt logic to Message
It _kinda_ makes more sense to have the `Message` be in charge of the
event prompt, since it is also responsible of the other pieces of content
(such as the body, subject, etc.)

However, this smells a bit, in particular the need to prefix the
`event_prompt` method with `recipient` (because I initially sent the
_sender's_ prompt, which is why tests matter and testing first is cool,
by the way).

Anyway, it starts to feel like "message" and "transmission" are 2 different
concepts that should be reflected in the models – the _message_ is what the
player sends the app, and the _transmission_ is the message plus extra
info, such as the event prompt.