Added a Transmission model
It's almost a decorator over a `Message`, but it avoids some confusion.

(I'm not too confortable bindly delegating _everything_ to an AR model,
though, because AR models do **a lot**.)

Note that we _could_ go even further and have a `TransmissionsMailer` mailer,
if we stick to the "the resource names and defines the mailer"
pseudo-convention, but I find it overkill, and it would require the `Transmission` model to be ActiveJob-serializable. And don't find the ROI
very high for this work.

Now, the main issue with the current `Transmission` model are the 2 code
smells: the model needs to ask the _Game_ object for the `Round` at the
time of the transmission, so that it can _then_ ask it for the relevant
_event_. It looks more and more like `Message` and `Round` should be
associated…