Added Player-aware "accessors" to Round
So, a `Round` model stores a lot of data that is related either to the
Earther or the Explorer: `#earther_hand`, `#earther_event`, `#explorer_hand`, etc.

This makes for a cumbersome API when, having a `Player` object, one needs
to ask for its associated data: the _role_ played by the `Player` in the
current game must first be determined, and then used to meta-program the
name of the relevant method to call on the `Round` object.

Now, `Round#event_for` and `Round#hand_of` give access to the same data
in a more fiendly way: the `Round` deals with the role-determing stuff.