MessagesMailer#transmission requires a `Round` arg
Since the transmission is round-dependant (because it requires data under
a `Round` responsiblity, such as the recipient's event for the prompt),
I find it best to explicitly pass the relevant `Round` object instead
of presuming that it will always be the "current" round. After all, by
the time the transmission is sent, the "current" round could change, and
then the transmission would use the wrong data.

Of course, if `Message`s and `Round`s were connected, such a thing would
not be necessary…