PSA: it's never too late to turn around if you're down the wrong road
There is a bug in the transmission email, did you spot it? The 'role' event should use different tables wether the receiver is the Earther of the Explorer, but the code doesn't take this role into account (the event lookup is done via the i18n mechanism, and the translations aren't scoped by role in any way).

This is yet another problem caused by the confusion between played and role, and the lack of an actual, simple and idiot-proof way to determine which role a player plays in a given game. Let's bite the bullet and revise our models structure.