Added a Character model between Game and Player
This adds a bit of confusion, because both the character and the player
can have a name (this is a part of the game that we'd skipped over so
far), and because sometimes we want to reference the player (as the
_actual_ recipient of a transmission, for example) and sometimes the
character (as the pretend sender of said transmission).

We'll see if this makes things too confusing, but in the meantime,
the _role_ a character has in a given game is clearly defined, now.