Fixed: the event prompt if for the wrong character
This HAD to happen! I mixed up the prompts to show – it is the
RECIPIENT's event, not the sender's.

Since the recipient is already a "derived" info of `Message`, having
a proxy method to pretend that we're not offending Demether seems a bit
hypocritical. Besides, since a `Transmission` is little more than a
wrapper around a `Message`, it's not _really_ breaking the law
to chain `message.recipient.role`, is it?