Fixed mailbox tests
This was a suble bug, introduced when adding checks on the
acceptance of a message for a given game:

When calling `game.messages#<<`, an exception is raised if the
message is invalid. In the mailbox, this exception is used as a kind
of control flow, to bounce the email.

However, there are _other_ reasons for which a mail could bounce (wrong
game id, for example). But the tests that cover these reasons have been
written _before_ the checks were added, and the test messages _are_
invalid

Therefore, we have tests that pass (the mail bounces), but for the
wrong reason (because the message in the mail is invalid in the first
place).

I only caught this when writting a _new_ test, using the same setup
as the first tests, and making sure that it failed _for the excpected
reason_ before coding the implementation. TDDing saved my bacon on this
one!