Bouce messages for ended games
This forces me to had a lot of fixtures, because the setup for an
ended game is heavy (we need at minimum 5 rounds to reach the end
of a game events table). Or maybe I should have added fixtures for
a game ended by a player, and not because the end of an event table
has been reached?