Adding fixtures for a "running game"
Since the 1st round is special, we'll need fixtures for games
at this round, and _also_ at another round – but not the second-to-last.

To avoid any confusions, now that we have fixtures for games at 3 different
stages (freshly started, running, and close to finishing), I've renamed
the first to something more descriptive.