Fixed: excessive message distortion
A tricky bug that I could only catch by actually playing the game…

Basically, the distortion "level" was computed using _all the events_
drawn at the time of the transmission. Which means that when transmitting
the _last message of a given round_, because a new round is automatically
created beforehand, the distortion was excessive: the event of the newly
created round was taken into account, even though the transmitted
message had nothing to do with it.