Fixed the event selection for a transmission
This logic has been confusing me for too long. It's time to lay it out
clearly.

A **transmission** is the sending of a message, during a given round,
along with the event that the **recipient** can use for their reply.
Therefore, the event is from:

- The **current** round if the recipient is the **earther** (because the
earther _replies_ in the current round).
- The **next** round if the recipient is the **explorer** (because the
explorer's reply marks the start of the new round).