Integration test for a full (but short) game
This is the test that was in progress during recent commits, and
actually drove them. It's a bit brittle, like all integration tests,
but has already proven useful to identify missing features (see the
aforementioned previous commits).

Note that integration tests with emails as a UI require a bit of
boilerplate. Note also that the randomness of the game had to be
replaced with stubbing, which lead to interesting hacks.