Added a welcome page
Yes, it's a WEB page! This should be the only web UI of the game.

(The name of the resource is a clever play on the name of the default
controller/page in Rails (in dev), `WelcomeController`. I'm so funny.)

There is more CSS that stricly needed (including a whole color palette)
because having HTML emails and/or other pages (for debug purposes, or a
leaderboard for example) is not _totally_ out of the question.