WIP: starting games by email
We want to allow players (existing and new) to start a game by sending
“command” emails to the system. This means that a new mailbox should
process a well-formated email to create a new pending game, or update
and start an existing one. Creations and updates should also handle the
associated records, such as `Player` and `Character`.

(The idea is that each player has to write to the system, cc'ing the
other player, and put the name and role of their character in the
subject line.)

However, this feature reveals a problem with the way the models are
organized: we want to have pending games that don't have all the
required _characters_ (but the required players), and this is not
allowed at the moment.