Finished: starting games by email
This requires that the 2 players send an e-mail to `new_game@…`,
with the other player in CC, and the name and role of their character
in the subject line.

The first email received will prepare the game, creating the `Player`
records if needed. The second email will finish the preparation, and
start the game.

Note that, when such an email is received, the sender's name is used to
update their player name. Note also that, because the other player's
_character_ name is unknown when the email is received, we have to allow
the creation of `Character` record without a `:name` attribute. This,
in turn, added validation logic to `Game` and `Character`, with a
special context: `:game_start`.