Test: transmitting a valid message
This introduces 3 new things:

*   The incoming mail fixtures (which were easy to generate thanks with
    the conductor UI at
    `http://localhost:3000/rails/conductor/action_mailbox/inbound_emails`
*   The necessity to manually include `ActionMailer::TestHelper` in the
    test case in order to test the _sending_ of emails when processing
    _incoming_ emails.
*   The idea that the _messages_ sent in a given game should be kept.

This means that the processing of a message mail should have 2 effects:
a new `Message` record should be created, and a new mail should be sent
to transmit the message to its actual recipient. This makes our test
more of an _integration_ test. Which is fine by me.