Added the Message model
Quite minimal for now. However, there is one tricky thing in this commit.

Because I've chosen to use the "first message" as a `Message` fixture,
it is now implied that if all fixtures are loaded, the _email_ in which
this message was sent has already been processed. Which means that the
"first message" cannot be used for a mailbox test. (In other words: the
`first_message.eml` fixture is a bit useless now.)

I should probably add a second incoming mail fixture, one for which
there is no corresponding, post-processing `Message` yet. In the
meantime, I'll skip loading the messages fixtures in the mailbox tests.